+++ src/components/NavTab/index.js
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React, { useState, useEffect } from "react";
 import './style.css'
 import { NavTabItem } from '../NavTabItem'
 import { NavTabContent } from "../NavTabContent";
@@ -6,9 +6,28 @@ import { ProdutosTab } from "../TabsContent/ProdutosTab"
 import { UsuariosTab } from "../TabsContent/UsuariosTab"
 
 export const NavTab = () => {
+  const token = JSON.parse(sessionStorage.getItem('@token'));
+
+  const [produtos, setProdutos] = useState();
+
+  const getData = async () => {
+    await fetch('http://localhost:3333/cliente/all', {
+      method: 'get',
+      headers: {
+        authorization: `Bearer ${token}`
+      }
+    })
+      .then(response => response.json())
+      .then(json => {
+        setProdutos(json);
+      })
+  };
+
+  useEffect(() => {
+    getData();
+  }, []);
 
   const [activeTab, setActiveTab] = useState("produtos");
-
   return (
     <div className="nav-tab-div">
       <ul className="nav">
@@ -21,7 +40,7 @@ export const NavTab = () => {
       </ul>
       <div className="outlet">
         <NavTabContent id="produtos" activeTab={activeTab}>
-          <ProdutosTab />
+          <ProdutosTab produtos={produtos !== undefined ? produtos : []}/>
         </NavTabContent>
         <NavTabContent id="usuarios" activeTab={activeTab}>
           <UsuariosTab />
